Added field TreeNode.id and made necessary modifications to make everything compile afterwards
A reference to DocTree.node_count is passed to inline parser function, so it might be utilized by the new
TreeNode::new_from_id_ref constructor, which also increments the counter.
This is a placeholder solution, until a more satisfying one is found.

Also, started working on a few DocTree methods for adding nodes to the contained zipper.