Added a new TreeNodeType::Class
Docutils would use the Pending node to give classes to the children
of this node during the tree transformation stage, but it seems clearer
to simply implement a specific node for this purpose.

There are other types of "pending" elements as well,
such as the table of contents node. A separete node
will most likely be generated for it as well.